+++ src/components/bluepad32/include/parser/uni_hid_parser.h
@@ -39,7 +39,7 @@ typedef void (*report_parse_feature_report_fn_t)(struct uni_hid_device_s* d,
 typedef void (*report_set_player_leds_fn_t)(struct uni_hid_device_s* d, uint8_t leds);
 typedef void (*report_set_lightbar_color_fn_t)(struct uni_hid_device_s* d, uint8_t r, uint8_t g, uint8_t b);
 // start_delay_ms: a delayed start measured in milliseconds. Use 0 to start rumble immediately.
-// duration_ms: duration of rumble in milliseconds. The controller might limit the max duration.
+// duration_ms: duration of rumble in milliseconds. The controller might limit the max duration. 0 means stop rumble.
 // weak_magnitude: The magnitude for the "weak motor".
 // strong_magnitude: The magnitude for the "strong motor".
 // If the controller has only one motor, then the max value between "weak" and "strong" is used.